Initial 1.21.30 commit (unfinished)

+++ src/main/java/cn/nukkit/network/protocol/EmotePacket.java
@@ -33,7 +33,7 @@ public class EmotePacket extends DataPacket {
         this.emoteID = byteBuf.readString();
         this.xuid = byteBuf.readString();
         this.platformId = byteBuf.readString();
-        this.flags = (byte) byteBuf.readByte();
+        this.flags = byteBuf.readByte();
     }
 
     @Override
